fix: remove unnecessary format option from gdal_translate commands

+++ regions/nl/1_fetch.sh
@@ -11,8 +11,9 @@ jq -r '.features[] | .properties.file' features.geojson | parallel --eta --bar -
   set -e
   URL={}
   ID=$(basename "$URL" .tif)
+  [ -f "$DATA/tiles/$ID.jp2" ] && exit 0
   curl -so "$ID.tif" "$URL"
-  gdal_translate -q "$ID.tif" "$ID.jp2"
+  gdal_translate --quiet "$ID.tif" "$ID.jp2"
   mv "$ID.jp2" "$DATA/tiles/"
   rm -f $ID*
 '
